+++ openstack/css/v1/clusters/Delete.go
@@ -2,8 +2,8 @@ package clusters
 
 import "github.com/opentelekomcloud/gophertelekomcloud"
 
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	raw, err := client.Delete(client.ServiceURL("clusters", id), &golangsdk.RequestOpts{
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("clusters", id), &golangsdk.RequestOpts{
 		OkCodes: []int{200},
 		MoreHeaders: map[string]string{
 			"Content-Type": "application/json",
