+++ openstack/css/v1/clusters/Get.go
@@ -1,8 +1,17 @@
 package clusters
 
-import "github.com/opentelekomcloud/gophertelekomcloud"
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
+)
 
-func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
+func Get(client *golangsdk.ServiceClient, id string) (*Cluster, error) {
 	raw, err := client.Get(client.ServiceURL("clusters", id), nil, nil)
-	return
+	if err != nil {
+		return nil, err
+	}
+
+	var res Cluster
+	err = extract.Into(raw.Body, &res)
+	return &res, err
 }
