+++ openstack/css/v1/clusters/List.go
@@ -2,11 +2,16 @@ package clusters
 
 import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
 )
 
-func List(client *golangsdk.ServiceClient) (p pagination.Pager) {
-	return pagination.NewPager(client, client.ServiceURL("clusters"), func(r pagination.PageResult) pagination.Page {
-		return ClusterPage{pagination.SinglePageBase(r)}
-	})
+func List(client *golangsdk.ServiceClient) ([]Cluster, error) {
+	raw, err := client.Get(client.ServiceURL("clusters"), nil, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	var res []Cluster
+	err = extract.IntoSlicePtr(raw.Body, &res, "clusters")
+	return res, err
 }
